- Fixed a strange change in Favorite.php - Add Product.php, which has disappeared from main
